fix node impl
