+++ crates/maple_3d/src/nodes/directional_light.rs
@@ -105,14 +105,6 @@ impl Node for DirectionalLight {
         &mut self.transform
     }
 
-    fn get_children(&self) -> &Scene {
-        &self.children
-    }
-
-    fn get_children_mut(&mut self) -> &mut Scene {
-        &mut self.children
-    }
-
     fn get_events(&mut self) -> &mut EventReceiver {
         &mut self.events
     }
