+++ crates/maple_3d/src/nodes/point_light.rs
@@ -96,14 +96,6 @@ impl Node for PointLight {
         &mut self.transform
     }
 
-    fn get_children(&self) -> &Scene {
-        &self.children
-    }
-
-    fn get_children_mut(&mut self) -> &mut Scene {
-        &mut self.children
-    }
-
     fn get_events(&mut self) -> &mut EventReceiver {
         &mut self.events
     }
