+++ crates/maple_engine/src/nodes/container.rs
@@ -67,14 +67,6 @@ where
         &mut self.transform
     }
 
-    fn get_children(&self) -> &Scene {
-        &self.children
-    }
-
-    fn get_children_mut(&mut self) -> &mut Scene {
-        &mut self.children
-    }
-
     fn get_events(&mut self) -> &mut crate::components::EventReceiver {
         &mut self.events
     }
