+++ crates/maple_engine/src/nodes/empty.rs
@@ -29,16 +29,8 @@ impl Node for Empty {
         &mut self.transform
     }
 
-    fn get_children(&self) -> &Scene {
-        &self.children
-    }
-
     fn get_events(&mut self) -> &mut crate::components::EventReceiver {
         &mut self.events
-    }
-
-    fn get_children_mut(&mut self) -> &mut Scene {
-        &mut self.children
     }
 }
