+++ crates/maple_engine/src/scene/mod.rs
@@ -1,5 +1,5 @@
 mod core;
-mod new_scene;
+// mod new_scene;
 mod scene_builder;
 
 pub use core::Scene;
