+++ crates/maple_physics/src/nodes/collider.rs
@@ -68,14 +68,6 @@ impl Node for Collider3D {
         &mut self.transform
     }
 
-    fn get_children_mut(&mut self) -> &mut Scene {
-        &mut self.children
-    }
-
-    fn get_children(&self) -> &Scene {
-        &self.children
-    }
-
     fn get_events(&mut self) -> &mut EventReceiver {
         &mut self.events
     }
