+++ crates/maple_physics/src/nodes/rigid_body.rs
@@ -44,14 +44,6 @@ impl Node for RigidBody3D {
     fn get_transform(&mut self) -> &mut NodeTransform {
         &mut self.transform
     }
-
-    fn get_children(&self) -> &Scene {
-        &self.children
-    }
-
-    fn get_children_mut(&mut self) -> &mut Scene {
-        &mut self.children
-    }
 }
 
 impl RigidBody3D {
